FE-Refactor: getUsers 관련 함수 utils 로 분리
